refactor(util): group filter-state helpers under util/filters/
Moves meetingFilters, signageFilters, and tagOrder into util/filters/ --
the calendar Day/Week filter builder, the signage kiosk URL-param
filter parser, and the shared tag sort order they both use. Pure
relocation.
