+++ frontend/app/components/admin/export/ExportTab.tsx
@@ -9,7 +9,7 @@ import {
   SIGNAGE_MODE_TYPES,
   SIGNAGE_ROOM_SLUGS,
   SIGNAGE_ZOOM_SLUGS,
-} from "../../../../util/signageFilters";
+} from "../../../../util/filters/signageFilters";
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, CATEGORY_COLOR } from "../../../../util/rooms/filterColors";
 import type { ILeaseSettings } from "../../../../types/models";
 import { ALL_MEETING_EXPORT_FIELD_KEYS, type MeetingExportFieldKey } from "../../../../util/meetingExportFields";
