+++ frontend/app/components/atoms/TagList.tsx
@@ -1,6 +1,6 @@
 import React, { useLayoutEffect, useRef, useState } from "react";
 import styles from '../../../styles/components/atoms/TagList.module.scss';
-import { sortTags } from '../../../util/tagOrder';
+import { sortTags } from '../../../util/filters/tagOrder';
 import { MODE_ICON_SRC } from '../../../util/rooms/modeIcons';
 
 interface TagListProps {
