+++ frontend/app/components/calendar/desktop/CalendarSidebarShell.tsx
@@ -7,7 +7,7 @@ import IconButton from "../../atoms/IconButton";
 import { useSidebar } from "../../../context/SidebarContext";
 import { useBreakpoint } from "../../../../hooks/useBreakpoint";
 import { IMeeting } from "../../../../types/models";
-import { MeetingFilters } from "../../../../util/meetingFilters";
+import { MeetingFilters } from "../../../../util/filters/meetingFilters";
 
 interface CalendarSidebarShellProps {
   isLoggedIn: boolean | null;
