+++ frontend/app/components/calendar/desktop/CompactCalendarSidebar.tsx
@@ -3,7 +3,7 @@ import IconButton from '../../atoms/IconButton';
 import MiniCalendar from '../../atoms/MiniCalendar';
 import FilterGroup from '../../shared/FilterGroup';
 import { LOCATION_ITEMS, ZOOM_ITEMS, CALENDAR_ITEMS, MODE_ITEMS } from '../shared/MeetingsFilter';
-import { MeetingFilters } from '../../../../util/meetingFilters';
+import { MeetingFilters } from '../../../../util/filters/meetingFilters';
 import styles from '../../../../styles/components/calendar/desktop/CompactCalendarSidebar.module.scss';
 
 type FlyoutKey = 'calendar' | 'location' | 'video' | 'group';
