+++ frontend/app/components/calendar/shared/MeetingsFilter.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import FilterGroup, { FilterGroupItem } from '../../shared/FilterGroup';
 import { ROOM_COLORS, ZOOM_ROOM_COLOR, CATEGORY_COLOR } from '../../../../util/rooms/filterColors';
-import { MeetingFilters } from '../../../../util/meetingFilters';
+import { MeetingFilters } from '../../../../util/filters/meetingFilters';
 
 interface MeetingsFilterProps {
     filters: MeetingFilters;
